gen: increase width of arguments/variables
avoid redudant width extension instructions

+++ src/generate.c
@@ -539,9 +539,9 @@ move_array_t generate_quiet(int64_t side) {
    return moves;
 }
 
-void add_piecewise(__uint128_t set, uint8_t from, move_array_t* moves) {
+void add_piecewise(__uint128_t set, uint64_t from, move_array_t* moves) {
    for (; set; set &= set - 1) {
-      uint8_t to = bsf_branchless(set);
+      uint64_t to = bsf_branchless(set);
       move_t move = { ._ = { from, to, board[from], board[to] } };
       moves->data[moves->count++] = move;
    }
@@ -549,8 +549,8 @@ void add_piecewise(__uint128_t set, uint8_t from, move_array_t* moves) {
 
 void add_shiftwise(__uint128_t set, int64_t shift, move_array_t* moves) {
    for (; set; set &= set - 1) {
-      uint8_t to = bsf_branchless(set);
-      uint8_t from = to - shift;
+      uint64_t to = bsf_branchless(set);
+      uint64_t from = to - shift;
       move_t move = { ._ = { from, to, board[from], board[to] } };
       moves->data[moves->count++] = move;
    }
